delete Question.js, all in Questionnaire.js
